refactor: consolidate specialized workspace modules into core database engine crate
