+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 members = [
     "crates/laminar-core",
-    "crates/laminar-ai",
     "crates/laminar-sql",
     "crates/laminar-connectors",
     "crates/laminar-server",
