--- crates/laminar-ai/Cargo.toml
@@ -1,68 +1,0 @@
-[package]
-name = "laminar-ai"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-documentation = "https://laminardb.io/api/laminar_ai"
-description = "Backend-agnostic AI inference for LaminarDB - model registry, provider trait, result cache, and local/remote backends"
-
-[dependencies]
-# Error handling
-thiserror = { workspace = true }
-
-# Async trait objects (dyn InferenceProvider)
-async-trait = { workspace = true }
-
-# Result cache (S3-FIFO in-memory), same crate the lookup/schema caches use.
-foyer = { workspace = true }
-
-# Content hashing for cache keys — xxh3-128 (not cryptographic; a fast,
-# collision-negligible key for a result cache, cheaper than SHA-256 near the
-# compute thread).
-xxhash-rust = { workspace = true }
-
-# Call-log buffer mutex (no poisoning; clippy bans std::sync::Mutex).
-parking_lot = { workspace = true }
-
-# --- remote backend (feature = "remote") ---
-# HTTP transport; matches the version laminar-connectors already compiles
-# (rustls, no openssl). serde(_json) for request/response bodies; tokio for
-# timeout/sleep; futures for bounded-concurrent per-row dispatch.
-reqwest = { version = "0.13", default-features = false, features = ["json", "rustls"], optional = true }
-serde = { workspace = true, optional = true }
-serde_json = { workspace = true, optional = true }
-tokio = { workspace = true, optional = true }
-futures = { workspace = true, optional = true }
-
-# Client-side rate limiting (token bucket / GCRA) for remote providers — the
-# standard maintained Rust limiter, rather than a hand-rolled bucket.
-governor = { version = "0.8", optional = true }
-
-# --- local backend (feature = "local") ---
-# ONNX Runtime via `ort`. tract (pure Rust) cannot analyse the dynamic
-# attention-mask broadcast in stock HuggingFace encoder exports; ort runs them
-# unchanged. `load-dynamic` loads `onnxruntime.dll`/`.so` at runtime instead of
-# static-linking the prebuilt C++ objects — so the build is independent of the
-# host toolchain (the static prebuilt needs a newer MSVC STL than some boxes
-# have) and the operator supplies the shared library (ORT >= 1.24, API 24).
-# tokenizers uses the pure-Rust fancy-regex backend (not onig/C); serde_json
-# reads config.json (id2label); tokio for spawn_blocking + the inference
-# deadline; reqwest fetches absent `hf:` repos from the HF CDN.
-ort = { version = "=2.0.0-rc.12", default-features = false, features = ["load-dynamic", "api-24"], optional = true }
-tokenizers = { version = "0.21", default-features = false, features = ["fancy-regex"], optional = true }
-
-[features]
-default = []
-# Remote LLM backends (OpenAI-compatible, Anthropic) over HTTP. Default build
-# carries zero ML/HTTP weight.
-remote = ["dep:reqwest", "dep:serde", "dep:serde_json", "dep:tokio", "dep:futures", "dep:governor"]
-# Local ONNX backend (ort/ONNX Runtime + HF tokenizers; reqwest for hf: fetch).
-local = ["dep:ort", "dep:tokenizers", "dep:serde_json", "dep:tokio", "dep:reqwest"]
-
-[dev-dependencies]
-tokio = { workspace = true }
-# Scratch model cache for the local-backend integration test.
-tempfile = { workspace = true }
