+++ crates/laminar-db/src/ai/mod.rs
