+++ crates/laminar-db/src/ai/registry.rs
