+++ crates/laminar-db/src/operator/ai_inference.rs
@@ -37,7 +37,7 @@ use rustc_hash::FxHashMap;
 use tokio::runtime::Handle;
 use tokio::sync::mpsc;
 
-use laminar_ai::{
+use crate::ai::{
     content_hash, params_version, AiCacheKey, AiCallLog, AiResultCache, BackendKind, CachedOutput,
     InferenceParams, InferenceProvider, Task,
 };
@@ -530,7 +530,7 @@ mod tests {
     use std::time::{Duration, Instant};
 
     use async_trait::async_trait;
-    use laminar_ai::{InferenceOutputs, InferenceRequest, InferenceResponse, ProviderError, Usage};
+    use crate::ai::{InferenceOutputs, InferenceRequest, InferenceResponse, ProviderError, Usage};
     use tokio::runtime::Handle;
 
     /// Provider that sleeps then echoes `L:<input>` for each row.
