+++ crates/laminar-server/Cargo.toml
@@ -19,10 +19,8 @@ path = "src/main.rs"
 # server only needs the unified facade plus laminar-core for things like
 # metrics types referenced directly.
 laminar-core = { path = "../laminar-core", version = "0.23.0" }
-laminar-db = { path = "../laminar-db", version = "0.23.0", features = ["api"] }
+laminar-db = { path = "../laminar-db", version = "0.23.0", features = ["api", "remote", "local"] }
 laminar-sql = { path = "../laminar-sql", version = "0.23.0" }
-# Remote + local AI backends for building the AI runtime from `[ai]`/`[models]`.
-laminar-ai = { path = "../laminar-ai", version = "0.23.0", features = ["remote", "local"] }
 
 # For pgwire dispatch on parsed AST
 sqlparser = { workspace = true }
